+++ src/simulator_core/entities/assets/controller/asset_controller_abstract.py
